# Performance optimization: Comprehensive frontend and nginx enhancements