--- a/docker-compose.yml
+++ b/docker-compose.yml
@@ -800,8 +800,8 @@ services:
       - GF_SECURITY_ADMIN_USER=admin
       - GF_SECURITY_ADMIN_PASSWORD=${GRAFANA_PASSWORD:-admin123}
       - GF_USERS_ALLOW_SIGN_UP=false
-      - GF_DOMAIN=grafana.ethical-ai-insider.com  
-  volumes:
+      - GF_DOMAIN=grafana.ethical-ai-insider.com
+    volumes:
       - grafana_data:/var/lib/grafana
       - ./monitoring/grafana/dashboards:/etc/grafana/provisioning/dashboards:ro
       - ./monitoring/grafana/datasources:/etc/grafana/provisioning/datasources:ro