--- a/frontend/package.json
+++ b/frontend/package.json
@@ -6,9 +6,12 @@
     "dev": "next dev -p 4000",
     "dev:staff": "next dev -p 4001",
     "build": "next build",
+    "build:analyze": "ANALYZE=true next build",
     "start": "next start",
     "start:staff": "next start -p 3001",
-    "lint": "next lint"
+    "lint": "next lint",
+    "bundle-analyzer": "npx @next/bundle-analyzer",
+    "performance-test": "node ../tools/performance-benchmark.js"
   },
   "dependencies": {
     "@headlessui/react": "^2.2.6",
@@ -43,6 +46,7 @@
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
+    "critters": "^0.0.23",
     "eslint": "^9",
     "eslint-config-next": "15.4.5",
     "tailwindcss": "^4",