… monitoring/alerts/frontend-slo-rules.yml
… nginx/prod-optimized.conf
… tools/performance-benchmark.js